ERT-1059 Reorganization of some widgets
