+++ devel/python/python/ert_gui/ertwidgets/caselist.py
@@ -1,11 +1,60 @@
+from PyQt4.QtCore import QSize
 from PyQt4.QtGui import QListWidget, QMessageBox, QAbstractItemView, QWidget, QVBoxLayout, QLabel
+from PyQt4.QtGui import QToolButton, QHBoxLayout
 
 from ert_gui import ERT
 from ert_gui.ertwidgets import addHelpToWidget
 from ert_gui.ertwidgets.models.ertmodel import getAllCases, selectOrCreateNewCase
 from ert_gui.ertwidgets.validateddialog import ValidatedDialog
-from ert_gui.widgets.add_remove_widget import AddRemoveWidget
+from ert_gui.widgets.util import resourceIcon
 
+
+class AddRemoveWidget(QWidget):
+    """
+    A simple class that provides to vertically positioned buttons for adding and removing something.
+    The addFunction and removeFunction functions must be provided.
+    """
+
+    def __init__(self, addFunction=None, removeFunction=None, horizontal=False):
+        QWidget.__init__(self)
+
+        self.addButton = QToolButton(self)
+        self.addButton.setIcon(resourceIcon("add"))
+        self.addButton.setIconSize(QSize(16, 16))
+        self.addButton.clicked.connect(addFunction)
+
+        self.removeButton = QToolButton(self)
+        self.removeButton.setIcon(resourceIcon("remove"))
+        self.removeButton.setIconSize(QSize(16, 16))
+        self.removeButton.clicked.connect(removeFunction)
+
+        if horizontal:
+            self.buttonLayout = QHBoxLayout()
+        else:
+            self.buttonLayout = QVBoxLayout()
+
+        self.buttonLayout.setMargin(0)
+
+        if horizontal:
+            self.buttonLayout.addStretch(1)
+
+        self.buttonLayout.addWidget(self.addButton)
+        self.buttonLayout.addWidget(self.removeButton)
+
+        if not horizontal:
+            self.buttonLayout.addStretch(1)
+        else:
+            self.buttonLayout.addSpacing(2)
+
+        self.setLayout(self.buttonLayout)
+
+    def enableAddButton(self, state):
+        """Enable or disable the add button"""
+        self.addButton.setEnabled(state)
+
+    def enableRemoveButton(self, state):
+        """Enable or disable the remove button"""
+        self.removeButton.setEnabled(state)
 
 class CaseList(QWidget):
