+++ devel/python/python/ert_gui/simulation/__init__.py
@@ -1,7 +1,9 @@
+from .progress import Progress
+from .simple_progress import SimpleProgress
+from .run_dialog import RunDialog
 from .simulation_config_panel import SimulationConfigPanel
 from .ensemble_experiment_panel import EnsembleExperimentPanel
 from .ensemble_smoother_panel import EnsembleSmootherPanel
 from .iterated_ensemble_smoother_panel import IteratedEnsembleSmootherPanel
 from .multiple_data_assimilation_panel import MultipleDataAssimilationPanel
-from .run_dialog import RunDialog
