+++ devel/python/python/ert_gui/simulation/run_dialog.py
@@ -1,13 +1,13 @@
 from threading import Thread
-from PyQt4.QtCore import Qt, pyqtSignal, QTimer, QSize
+
+from PyQt4.QtCore import Qt, QTimer, QSize
 from PyQt4.QtGui import QDialog, QVBoxLayout, QLayout, QMessageBox, QPushButton, QHBoxLayout, QColor, QLabel
 
+from ert_gui.simulation import Progress, SimpleProgress
 from ert_gui.simulation.models import BaseRunModel, SimulationsTracker
 from ert_gui.tools.plot.plot_tool import PlotTool
 from ert_gui.widgets import util
 from ert_gui.widgets.legend import Legend
-from ert_gui.widgets.progress import Progress
-from ert_gui.widgets.simple_progress import SimpleProgress
 
 
 class RunDialog(QDialog):
