+++ devel/python/python/ert_gui/tools/plot/CMakeLists.txt
@@ -16,3 +16,4 @@ set(PYTHON_SOURCES
 add_python_package("python.ert_gui.tools.plot" ${PYTHON_INSTALL_PREFIX}/ert_gui/tools/plot "${PYTHON_SOURCES}" True)
 
 add_subdirectory(customize)
+add_subdirectory(widgets)
