+++ devel/python/python/ert_gui/tools/plot/widgets/CMakeLists.txt
@@ -1,0 +1,8 @@
+set(PYTHON_SOURCES
+    __init__.py
+    clearable_line_edit.py
+    custom_date_edit.py
+)
+
+add_python_package("python.ert_gui.tools.plot.widgets" ${PYTHON_INSTALL_PREFIX}/ert_gui/tools/plot/widgets "${PYTHON_SOURCES}" True)
+
