+++ devel/python/python/ert_gui/tools/plot/widgets/clearable_line_edit.py
