+++ devel/python/python/ert_gui/widgets/CMakeLists.txt
@@ -1,16 +1,11 @@
 set(PYTHON_SOURCES
     __init__.py
-    add_remove_widget.py
-    clearable_line_edit.py
-    custom_date_edit.py
     error_popup.py
     helped_widget.py
     legend.py
     path_chooser.py
     path_format.py
-    progress.py
     search_box.py
-    simple_progress.py
     string_box.py
     util.py
 )
